+++ CHANGELOG.md
@@ -9,6 +9,34 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [0.7.1] - 2024-03-05
+
+### Added
+
+- Implemented `From<Configuration>` for `dogstatsd::Options`
+
+### Changed
+
+- `Config` is now `Configuration`
+
+### Removed
+
+- The `Configuration` trait
+- `Country` and `Environment` serialization
+
+---
+
+## [0.7.0] - 2024-03-04
+
+### Changed
+
+- **BREAKING CHANGES**:
+  - `PrimaConfiguration` is now `Config`
+  - `Config::new` now takes only 2 arguments, `to_addr` and `namespace`
+  - `from_addr` is now `0.0.0.0:0` by default, but can be customized using `.with_from_addr`
+
+---
+
 ## [0.6.0] - 2023-10-02
 
 ### Added
@@ -19,7 +47,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
-- **BREAKING CHANGE**: `PrimaConfiguration::new` function no longer takes `env` parameter. 
+- **BREAKING CHANGE**: `PrimaConfiguration::new` function no longer takes `env` parameter.
   - Use `with_environment` function to manually set the `env:{env}` tag
   - If no environment is set the library will use `DD_ENV` var by default if set.
 
@@ -100,7 +128,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.5]
 
-[Unreleased]: https://github.com/primait/prima_datadog.rs/compare/0.6.0...HEAD
+[Unreleased]: https://github.com/primait/prima_datadog.rs/compare/0.7.1...HEAD
+[0.7.1]: https://github.com/primait/prima_datadog.rs/compare/0.7.0...0.7.1
+[0.7.0]: https://github.com/primait/prima_datadog.rs/compare/0.7.0...0.7.0
 [0.6.0]: https://github.com/primait/prima_datadog.rs/compare/0.5.0...0.6.0
 [0.5.0]: https://github.com/primait/prima_datadog.rs/compare/0.4.0...0.5.0
 [0.4.0]: https://github.com/primait/prima_datadog.rs/compare/0.3.1...0.4.0